Drop the TestWebEdit project that didn't actually test anything
Initially I thought it was a unit-test of some sort, but in fact it was
only an interactive testbed for the author to get familiar with some WinAPI
functions. I would have kept it if it was testing some library or wrapper
for those functions. Instead, when the knowledge was received the code was
copy-pasted inline into the plugin, so there is no need to keep the project
around anymore.