Import the missing legacy source files
These files were released in the v2.1 package, so they were always part of
the open-source distribution. However, the original Subversion repository
pulled them via the svn:externals dependency from a proprietary library.
Now that we are no longer working with that library, nor with Subversion,
we can simply copy the latest version of the files.