Build WebEdit into folders with portable Notepad++ for easy testing
Simply copy a portable distribution of Notepad++ to the
WebEdit\bin\npp.portable[.x64] directory, and the DLLs will be placed in
the correct target-specific subfolder on build. To run the corresponding
Notepad++ instance on F5 it's also necessary to configure the correct
"Start external program" in the Debug section of the project properties.