Fix compilation by removing the XC redirection file
For some reason XC can't find the Windows.sym file in its default path,
even if the default value "*.sym = .;$!/../sym/x86" is set (exactly as it
is in the default xc.red).

Remove the xc.red file and move the sources from ..\Lib to the main folder.
Update the make rules accordingly.